+++ src/messages/Message.hpp
@@ -61,6 +61,13 @@ struct Message {
     /// Map of extra data associated with each Twitch badge
     std::unordered_map<QString, QString> twitchBadgeInfos;
 
+    /// List of external badges associated with this message
+    /// The badge should follow the following format: "provider:badgename". e.g.:
+    ///  - betterttv:pro
+    ///  - frankerfacez:mod
+    ///  - 7tv:sub
+    QStringList externalBadges;
+
     std::shared_ptr<QColor> highlightColor;
     // Each reply holds a reference to the thread. When every reply is dropped,
     // the reply thread will be cleaned up by the TwitchChannel.