+++ src/providers/bttv/BttvBadges.cpp
@@ -3,6 +3,8 @@
 #include "messages/Emote.hpp"
 #include "messages/Image.hpp"
 
+using namespace Qt::StringLiterals;
+
 namespace chatterino {
 
 QString BttvBadges::idForBadge(const QJsonObject &badgeJson) const
@@ -11,10 +13,10 @@ QString BttvBadges::idForBadge(const QJsonObject &badgeJson) const
 }
 
 EmotePtr BttvBadges::createBadge(const QString &id,
-                                 const QJsonObject & /* badgeJson */) const
+                                 const QJsonObject & /*badgeJson*/) const
 {
     auto emote = Emote{
-        .name = EmoteName{},
+        .name = EmoteName{u"betterttv:pro"_s},
         .images = ImageSet(Image::fromUrl(Url{id}, 18.0 / 72.0)),
         .tooltip = Tooltip{"BTTV Pro"},
         .homePage = Url{},